Showcase i18n + middleware + edge runtime bug
The bug is that the locale in the context passed to  `getServerSideProps` is always the default locale on a fresh page load when using the edge runtime and in the presence of a middleware.

The bug appears for both static and dynamic routes.

The bug does not appear when clicking links within the app or when not using the edge runtime or when not using a middleware (either because one doesn't exist or because it doesn't match the current route).